修正前
