+++ include/env.h
@@ -2,8 +2,34 @@
 #ifndef ENV_H
 # define ENV_H
 
-# include "ft_hash.h"
+# define HASH_TABLE_SIZE 101 // 適当に決めたサイズ
 
-t_hash_table	*create_env_table(void);
+typedef struct s_hash_node
+{
+	char				*key;
+	char				*value;
+	struct s_hash_node	*next;
+}						t_hash_node;
+
+typedef struct s_hash_table
+{
+	t_hash_node			*table[HASH_TABLE_SIZE];
+	char				*(*search)(struct s_hash_table *table, const char *key);
+	int					(*delete)(struct s_hash_table *table, const char *key);
+	char				*(*insert)(struct s_hash_table *table, char *key,
+						char *value);
+	void				(*free)(struct s_hash_table *table);
+}						t_hash_table;
+
+# include "env.h"
+
+t_hash_table			*create_hash_table(void);
+
+char					*insert(t_hash_table *table, char *key, char *value);
+char					*search(t_hash_table *table, const char *key);
+int delete (t_hash_table *table, const char *key);
+void					free_hash_table(t_hash_table *table);
+
+t_hash_table			*create_env_table(void);
 
 #endif /* ENV_H */
